requirements.txt
# LiveKit Agent SDK
livekit-agents>=0.12.0
livekit-api>=0.7.0
livekit>=0.17.0

# LiveKit Plugins for STT/TTS/VAD
livekit-plugins-deepgram>=0.7.0
livekit-plugins-openai>=0.9.0
livekit-plugins-cartesia>=0.5.0
livekit-plugins-silero>=0.7.0

# HTTP/API
aiohttp>=3.9.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0

# Twilio (for calls + SMS)
twilio>=8.10.0

# Utilities
python-dotenv>=1.0.0
motor
pymongo

sinch>=1.1.4